test_training.py
#!/usr/bin/env python3
"""
Simple test script to validate the refactored ML pipeline works correctly.
"""

import json
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, accuracy_score
from nltk.stem import PorterStemmer
from nltk.tokenize import word_tokenize

def load_json_file(file_path):
    """Load a JSON file safely."""
    with open(file_path, 'r', encoding='utf-8') as f:
        return json.load(f)

def load_corpus():
    """Load training corpus from files."""
    texts = []
    intents = []
    
    training_files = [
        'data/processed/base-corpus.json',
        'data/processed/rhcp-corpus.json'
    ]
    
    for file_path in training_files:
        corpus = load_json_file(file_path)
        
        for item in corpus.get('data', []):
            if item.get('intent') and item.get('intent') != 'None':
                for utterance in item.get('utterances', []):
                    if utterance and utterance.strip():  # Skip empty utterances
                        texts.append(utterance.strip())
                        intents.append(item['intent'])
    
    return texts, intents

def main():
    """Run the basic training test."""
    print("🚀 RHCP Chatbot Training Test")
    print("=" * 40)
    
    try:
        # Load data
        print("📂 Loading training data...")
        texts, intents = load_corpus()
        df = pd.DataFrame({'text': texts, 'intent': intents})
        
        print(f"✅ Loaded {len(df)} samples with {df['intent'].nunique()} unique intents")
        
        # Analyze class distribution
        class_counts = df['intent'].value_counts()
        print(f"📊 Top 5 intents:")
        for i, (intent, count) in enumerate(class_counts.head().items()):
            print(f"  {i+1}. {intent}: {count} samples")
        
        imbalance_ratio = class_counts.iloc[0] / class_counts.iloc[-1]
        print(f"⚖️ Class imbalance ratio: {imbalance_ratio:.1f}:1")
        
        # Split data
        print("\n🔄 Splitting data...")
        X_train, X_test, y_train, y_test = train_test_split(
            df['text'], df['intent'], 
            test_size=0.2, 
            random_state=42, 
            stratify=df['intent']
        )
        print(f"✅ Train: {len(X_train)}, Test: {len(X_test)}")
        
        # Create simple pipeline without custom tokenizer (to avoid pickle issues)
        print("\n🤖 Creating model pipeline...")
        pipeline = Pipeline([
            ('tfidf', TfidfVectorizer(
                ngram_range=(1, 3),
                stop_words='english',
                max_features=5000
            )),
            ('clf', LogisticRegression(
                random_state=42,
                solver='lbfgs',
                max_iter=1000,
                class_weight='balanced'
            ))
        ])
        
        # Train model
        print("🎯 Training model...")
        pipeline.fit(X_train, y_train)
        print("✅ Training completed")
        
        # Evaluate
        print("\n📊 Evaluating model...")
        y_pred = pipeline.predict(X_test)
        accuracy = accuracy_score(y_test, y_pred)
        
        print(f"🎯 Test Accuracy: {accuracy:.4f}")
        
        # Test on specific cases
        print("\n🧪 Testing on specific cases...")
        test_cases = [
            'are you a bot',
            'bye for now',
            'Hello',
            'Who are the members of the band?',
            'Tell me about quantum physics',
            'when was RHCP formed'
        ]
        
        predictions = pipeline.predict(test_cases)
        probabilities = pipeline.predict_proba(test_cases)
        
        for i, (text, pred) in enumerate(zip(test_cases, predictions)):
            confidence = np.max(probabilities[i])
            print(f"  '{text}' -> '{pred}' (confidence: {confidence:.3f})")
        
        print(f"\n✅ Training test completed successfully!")
        print(f"📈 Model achieved {accuracy:.1%} accuracy on test set")
        
        if accuracy > 0.6:
            print("🎉 Good performance! The refactored pipeline is working correctly.")
        else:
            print("⚠️ Performance could be improved, but basic functionality is working.")
        
        return True
        
    except Exception as e:
        print(f"❌ Training test failed: {e}")
        import traceback
        traceback.print_exc()
        return False

if __name__ == "__main__":
    main() 