requirements_ml.txt
# RHCP Chatbot ML Pipeline Dependencies
# =====================================

# Core ML/Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.2.0

# Natural Language Processing
nltk>=3.8.0

# Configuration Management
PyYAML>=6.0

# Web Framework (existing)
fastapi>=0.104.1
uvicorn>=0.24.0

# Database (existing)
sqlalchemy>=2.0.23
pydantic>=2.5.0

# Testing (existing)
pytest>=7.4.3

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Additional utilities
pathlib2>=2.3.7; python_version < "3.6"
typing-extensions>=4.0.0

# Optional: Advanced ML (uncomment if needed)
# xgboost>=1.7.0
# lightgbm>=3.3.0
# optuna>=3.0.0  # for hyperparameter optimization 